+++ samples/sampler/stories/default/Heading.js
@@ -1,9 +1,9 @@
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import {boolean, select, text} from '@enact/storybook-utils/addons/knobs';
+import BodyText from '@enact/ui/BodyText';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 
-import BodyText from '@enact/ui/BodyText';
 import Heading, {HeadingBase} from '@enact/agate/Heading';
 
 import css from './Heading.module.less';
@@ -25,6 +25,7 @@ storiesOf('Agate', module)
 			const knobProps = {
 				color: select('color', prop.colors, Config),
 				marqueeOn: select('marqueeOn', prop.marqueeOn, Config),
+				showBackButton: boolean('showBackButton', Config),
 				showLine: boolean('showLine', Config),
 				size: select('size', prop.sizes, Config),
 				spacing: select('spacing', prop.spacings, Config)